refactor: flatten directory structure
