+++ src/angular/table/public-api.ts
@@ -3,9 +3,9 @@ export * from './table.module';
 export * from './table/table-data-source';
 export * from './table/table';
 export * from './sort/sort-direction';
-export * from './sort-header/sort-header';
+export * from './sort/sort-header';
 export * from './sort/sort';
-export * from './sort-header/sort-animations';
-export * from './text-column/text-column';
-export * from './table-cell/table-cell';
-export * from './table-row/table-row';
+export * from './sort/sort-animations';
+export * from './table/text-column';
+export * from './table/cell';
+export * from './table/row';
